Chakinzo's branch

+++ Shoot The Komsai/bossKomsai.h
@@ -3,22 +3,22 @@
 
 class BossKomsai {
 private:
-    int bossKomsaiX, bossKomsaiY;
-    int speedX, speedY;  // horizontal & vertical speed
+    float bossKomsaiX, bossKomsaiY;
+    float speedX, speedY; 
     int bossHealth;
 
 public:
     BossKomsai();
 
-    void set_BossKomsaiX(int val);
-    void set_BossKomsaiY(int val);
+    void set_BossKomsaiX(float val);
+    void set_BossKomsaiY(float val);
     int get_BossKomsaiX() const;
     int get_BossKomsaiY() const;
 
-    void set_BossSpeedX(int val);
-    void set_BossSpeedY(int val);
-    int get_BossSpeedX() const;
-    int get_BossSpeedY() const;
+    void set_BossSpeedX(float val);
+    void set_BossSpeedY(float val);
+    float get_BossSpeedX() const;
+    float get_BossSpeedY() const;
 
     void decrement_BossHealth(); 
     int get_BossHealth() const;
